requirements.txt
Cython==0.29.32
librosa==0.9.2
matplotlib==3.5.3
numpy==1.23.3
phonemizer==3.2.1
scipy==1.9.1
Unidecode==1.3.4
pytorch-lightning==1.7.5

pypinyin==0.47.1